+++ tests/asserter.js
@@ -1,13 +1,14 @@
 require('..')(module);
 
 var util = require('util'),
+    Generator = require('../lib/asserter/generator'),
     testedClass = require('../lib/asserter'),
     unit = module.exports = {
         testClass: function() {
             var object, generator;
 
             this
-                .if(generator = {})
+                .if(generator = new Generator())
                 .then()
                     .object(object = new testedClass(generator))
                     .undefined(object.value)
@@ -19,7 +20,7 @@ var util = require('util'),
             var object;
 
             this
-                .if(object = new testedClass({}))
+                .if(object = new testedClass(new Generator()))
                 .then()
                     .object(object.setWith()).isIdenticalTo(object)
             ;
@@ -29,7 +30,7 @@ var util = require('util'),
             var object, message, param, otherParam;
 
             this
-                .if(object = new testedClass({}))
+                .if(object = new testedClass(new Generator()))
                 .and(message = Math.random().toString(36).substring(7))
                 .then()
                     .error(function() {
@@ -53,7 +54,7 @@ var util = require('util'),
             var object, generator;
 
             this
-                .if(generator = {})
+                .if(generator = new Generator())
                 .and(object = new testedClass(generator))
                 .then()
                     .object(object.check()).isIdenticalTo(object)